Vaporock Basalt notebook
A Jupyter notebook used to calculate the vapor and liquid species of a Basalt composition from 1500 to 3000K and from fo2=-4 to fo2=7. The densities', molar fractions and partial pressures are also plotted.
Outputs are saved as excel file, to organise this have this notebook in a directory called Basalt and then your excel files can be written into that directory.