BSE Vaporock notebook
A notebook to calculate the liquid and vapor species of a BSE composition for a range of temperatures and fo2 values using VapoRock.  The densities, molar fractions and partial pressures are also plotted.
Outputs are saved as excel file, to organise this have this notebook in a directory called BSEand then your excel files can be written into that directory.