VapoRock BSE notebook
A notebook to calculate the liquid and vapor species of a BSE composition for a range of temperatures (1500-3000) and fo2 (-4 to 7) values using VapoRock.  The densities, molar fractions and partial pressures are also plotted.
Outputs are saved as excel file, to organise this have this notebook in a directory called BSE and then your excel files can be written into that directory.